+++ actix-web/src/guard/mod.rs
@@ -110,6 +110,12 @@ impl<'a> GuardContext<'a> {
     pub fn header<H: Header>(&self) -> Option<H> {
         H::parse(self.req).ok()
     }
+
+    /// Counterpart to [HttpRequest::app_data](crate::HttpRequest::app_data).
+    #[inline]
+    pub fn app_data<T: 'static>(&self) -> Option<&T> {
+        self.req.app_data()
+    }
 }
 
 /// Interface for routing guards.
@@ -511,5 +517,19 @@ mod tests {
             .method(Method::TRACE)
             .to_srv_request();
         assert!(guard.check(&req.guard_ctx()));
+    }
+
+    #[test]
+    fn app_data() {
+        const TEST_VALUE: u32 = 42;
+        let guard = fn_guard(|ctx| dbg!(ctx.app_data::<u32>()) == Some(&TEST_VALUE));
+
+        let req = TestRequest::default().app_data(TEST_VALUE).to_srv_request();
+        assert!(guard.check(&req.guard_ctx()));
+
+        let req = TestRequest::default()
+            .app_data(TEST_VALUE * 2)
+            .to_srv_request();
+        assert!(!guard.check(&req.guard_ctx()));
     }
 }