+++ actix-web/CHANGES.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Add `guard::GuardContext::app_data()` method.
 - Implement `From<Box<dyn ResponseError>>` for `Error`.
 - Add `CustomizeResponder::add_cookie()` method.